requirements.txt
PyPDF2==3.0.1
langchain
langchain-perplexity
langchain_community
python-dotenv
streamlit
faiss-cpu
fastembed